Refactor code for loading and leaderboard display
